requirements.txt
blue==0.9.1
celery==5.3.1
Django==4.2.2
djangorestframework==3.14.0
isort==5.12.0
pip-api==0.0.30
pytest==7.3.2
gunicorn==20.1.0
django-celery-results==2.5.1
django-cors-headers==4.1.0
whitenoise==6.4.0
psycopg==3.1.9
